rename: query 도메인 폴더로 분리
