+++ src/services/queries/upload-revenue/useRevenueUploadHistory.ts
@@ -5,7 +5,7 @@ import {
 import { MOCK_TRANSACTION_UPLOAD } from "@/constants/mock";
 import { MEMBER_TYPE } from "@/types/enums/user.enum";
 
-import { IGETTransactionUploadResponse } from "../api/types/transaction";
+import { IGETTransactionUploadResponse } from "../../api/types/transaction";
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const getRevenueUploadHistory = (month: string): Promise<IGETTransactionUploadResponse> => {
